requirements.txt
aiogram==3.6.0
vk_api
python-dotenv